+++ jaguar_jwt/CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 
-+ 2.1.2
+## TBD
+
++ Added support for optional Not Before (`nbf`) time claims.
++ Fixed validation to reject token when current time equals the Expiry time.
+
+## 2.1.2
 
 + Fixed when `typ` is not present